Improve parameter logging: show actual values of array-typed parameters, fixes #5342
inspired by: LogValuesFormatter from Microsoft.Extensions.Logging.Abstractions
