+++ src/Npgsql/NpgsqlCommand.cs
@@ -17,6 +17,7 @@ using System.Threading.Channels;
 using Microsoft.Extensions.Logging;
 using Npgsql.Internal;
 using Npgsql.Properties;
+using System.Collections;
 
 namespace Npgsql;
 
@@ -1812,7 +1813,7 @@ GROUP BY pg_proc.proargnames, pg_proc.proargtypes, pg_proc.proallargtypes, pg_pr
                     LogMessages.ExecutingCommandWithParameters(
                         logger,
                         singleCommand.FinalCommandText!,
-                        ParametersDbNullAsString(singleCommand),
+                        GetParametersForLogging(singleCommand),
                         connector.Id);
                 }
                 else
@@ -1820,7 +1821,7 @@ GROUP BY pg_proc.proargnames, pg_proc.proargtypes, pg_proc.proallargtypes, pg_pr
                     LogMessages.CommandExecutionCompletedWithParameters(
                         logger,
                         singleCommand.FinalCommandText!,
-                        ParametersDbNullAsString(singleCommand),
+                        GetParametersForLogging(singleCommand),
                         connector.QueryLogStopWatch.ElapsedMilliseconds,
                         connector.Id);
                 }
@@ -1839,7 +1840,7 @@ GROUP BY pg_proc.proargnames, pg_proc.proargtypes, pg_proc.proallargtypes, pg_pr
             {
                 var commands = new (string, object[])[InternalBatchCommands.Count];
                 for (var i = 0; i < InternalBatchCommands.Count; i++)
-                    commands[i] = (InternalBatchCommands[i].FinalCommandText!, ParametersDbNullAsString(InternalBatchCommands[i]));
+                    commands[i] = (InternalBatchCommands[i].FinalCommandText!, GetParametersForLogging(InternalBatchCommands[i]));
 
                 if (executing)
                     LogMessages.ExecutingBatchWithParameters(logger, commands, connector.Id);
@@ -1858,15 +1859,53 @@ GROUP BY pg_proc.proargnames, pg_proc.proargtypes, pg_proc.proallargtypes, pg_pr
             }
         }
 
-        object[] ParametersDbNullAsString(NpgsqlBatchCommand c)
+        static object[] GetParametersForLogging(NpgsqlBatchCommand c)
         {
             var positionalParameters = c.CurrentParametersReadOnly;
             var parameters = new object[positionalParameters.Count];
             for (var i = 0; i < positionalParameters.Count; i++)
-                parameters[i] = positionalParameters[i].Value == DBNull.Value ? "NULL" : positionalParameters[i].Value!;
+            {
+                parameters[i] = GetParameterForLogging(positionalParameters[i].Value);
+            }
             return parameters;
+
+            object GetParameterForLogging(object? value)
+            {
+                return value switch
+                {
+                    DBNull or null => "NULL",
+                    IEnumerable enumerable and not string => GetEnumerableForLogging(enumerable),
+                    _ => value
+                };
+
+                string GetEnumerableForLogging(IEnumerable enumerable)
+                {
+                    var vsb = new StringBuilder(256);
+                    var count = 0;
+                    vsb.Append('[');
+                    foreach (var e in enumerable)
+                    {
+                        if (count > 9)
+                        {
+                            vsb.Append(", ...");
+                            break;
+                        }
+
+                        if (count > 0)
+                        {
+                            vsb.Append(", ");
+                        }
+
+                        vsb.Append(GetParameterForLogging(e));
+                        count++;
+                    }
+
+                    vsb.Append(']');
+                    return vsb.ToString();
+                }
+            }
         }
-    }
+   }
 
     /// <summary>
     /// Create a new command based on this one.
