+++ test/Npgsql.Tests/BatchTests.cs
@@ -797,100 +797,6 @@ LANGUAGE 'plpgsql'");
 
     #endregion Miscellaneous
 
-    #region Logging
-
-    [Test]
-    public async Task Log_ExecuteScalar_single_statement_without_parameters()
-    {
-        await using var dataSource = CreateLoggingDataSource(out var listLoggerProvider);
-        await using var conn = await dataSource.OpenConnectionAsync();
-        await using var cmd = new NpgsqlBatch(conn)
-        {
-            BatchCommands = { new("SELECT 1") }
-        };
-
-        using (listLoggerProvider.Record())
-        {
-            await cmd.ExecuteScalarAsync();
-        }
-
-        var executingCommandEvent = listLoggerProvider.Log.Single(l => l.Id == NpgsqlEventId.CommandExecutionCompleted);
-
-        Assert.That(executingCommandEvent.Message, Does.Contain("Command execution completed").And.Contains("SELECT 1"));
-        AssertLoggingStateContains(executingCommandEvent, "CommandText", "SELECT 1");
-        AssertLoggingStateDoesNotContain(executingCommandEvent, "Parameters");
-
-        if (!IsMultiplexing)
-            AssertLoggingStateContains(executingCommandEvent, "ConnectorId", conn.ProcessID);
-    }
-
-    [Test]
-    public async Task Log_ExecuteScalar_multiple_statements_with_parameters()
-    {
-        await using var dataSource = CreateLoggingDataSource(out var listLoggerProvider);
-        await using var conn = await dataSource.OpenConnectionAsync();
-        await using var batch = new NpgsqlBatch(conn)
-        {
-            BatchCommands =
-            {
-                new("SELECT $1") { Parameters = { new() { Value = 8 } } },
-                new("SELECT $1, 9") { Parameters = { new() { Value = 9 } } }
-            }
-        };
-
-        using (listLoggerProvider.Record())
-        {
-            await batch.ExecuteScalarAsync();
-        }
-
-        var executingCommandEvent = listLoggerProvider.Log.Single(l => l.Id == NpgsqlEventId.CommandExecutionCompleted);
-
-        // Note: the message formatter of Microsoft.Extensions.Logging doesn't seem to handle arrays inside tuples, so we get the
-        // following ugliness (https://github.com/dotnet/runtime/issues/63165). Serilog handles this fine.
-        Assert.That(executingCommandEvent.Message, Does.Contain("Batch execution completed").And.Contains("[(SELECT $1, System.Object[]), (SELECT $1, 9, System.Object[])]"));
-        AssertLoggingStateDoesNotContain(executingCommandEvent, "CommandText");
-        AssertLoggingStateDoesNotContain(executingCommandEvent, "Parameters");
-
-        if (!IsMultiplexing)
-            AssertLoggingStateContains(executingCommandEvent, "ConnectorId", conn.ProcessID);
-
-        var batchCommands = (IList<(string CommandText, object[] Parameters)>)AssertLoggingStateContains(executingCommandEvent, "BatchCommands");
-        Assert.That(batchCommands.Count, Is.EqualTo(2));
-        Assert.That(batchCommands[0].CommandText, Is.EqualTo("SELECT $1"));
-        Assert.That(batchCommands[0].Parameters[0], Is.EqualTo(8));
-        Assert.That(batchCommands[1].CommandText, Is.EqualTo("SELECT $1, 9"));
-        Assert.That(batchCommands[1].Parameters[0], Is.EqualTo(9));
-    }
-
-    [Test]
-    public async Task Log_ExecuteScalar_single_statement_with_parameter_logging_off()
-    {
-        await using var dataSource = CreateLoggingDataSource(out var listLoggerProvider, sensitiveDataLoggingEnabled: false);
-        await using var conn = await dataSource.OpenConnectionAsync();
-        await using var batch = new NpgsqlBatch(conn)
-        {
-            BatchCommands =
-            {
-                new("SELECT $1") { Parameters = { new() { Value = 8 } } },
-                new("SELECT $1, 9") { Parameters = { new() { Value = 9 } } }
-            }
-        };
-
-        using (listLoggerProvider.Record())
-        {
-            await batch.ExecuteScalarAsync();
-        }
-
-        var executingCommandEvent = listLoggerProvider.Log.Single(l => l.Id == NpgsqlEventId.CommandExecutionCompleted);
-        Assert.That(executingCommandEvent.Message, Does.Contain("Batch execution completed").And.Contains("[SELECT $1, SELECT $1, 9]"));
-        var batchCommands = (IList<string>)AssertLoggingStateContains(executingCommandEvent, "BatchCommands");
-        Assert.That(batchCommands.Count, Is.EqualTo(2));
-        Assert.That(batchCommands[0], Is.EqualTo("SELECT $1"));
-        Assert.That(batchCommands[1], Is.EqualTo("SELECT $1, 9"));
-    }
-
-    #endregion Logging
-
     #region Initialization / setup / teardown
 
     // ReSharper disable InconsistentNaming
